Simplify tests.
- Move into one file.
