+++ karma.conf.js
@@ -31,8 +31,9 @@ module.exports = function(config) {
 
     // list of files / patterns to load in the browser
     files: [
+      'tests/test-karma.js',
       {
-        pattern: 'tests/test-karma.js',
+        pattern: 'tests/test.js',
         watched: false, served: true, included: true
       }
     ],
