+++ package.json
@@ -7,7 +7,7 @@
     "build": "npm run build-webpack",
     "build-webpack": "webpack",
     "test": "npm run test-node",
-    "test-node": "cross-env NODE_ENV=test mocha --delay -t 30000 -A -R ${REPORTER:-spec} tests/test.js",
+    "test-node": "cross-env NODE_ENV=test mocha -t 30000 -A -R ${REPORTER:-spec} tests/test.js",
     "test-karma": "karma start",
     "lint": "eslint '*.js' 'lib/*.js' 'tests/*.js'"
   },
