+++ tests/test-karma.js
@@ -1,0 +1,8 @@
+/**
+ * Karma test support.
+ *
+ * Copyright (c) 2011-2019 Digital Bazaar, Inc. All rights reserved.
+ */
+// load polyfills
+// no polyfills are currently required.
+
